+++ connector/connect/common/src/main/scala/org/apache/spark/sql/connect/client/SparkResult.scala
